loggly

+++ app/Http/Controllers/PostsController.php
@@ -6,9 +6,13 @@ use Illuminate\Http\Request;
 use App\Post;
 use DB;
 use Illuminate\Support\Facades\Storage;
+use Zizaco\Entrust\Traits\EntrustUserTrait;
+use App\User;
+
 
 class PostsController extends Controller
 {
+    use EntrustUserTrait;
     /**
      * Create a new controller instance.
      *
@@ -16,7 +20,6 @@ class PostsController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('auth', ['except' => ['index', 'show']]);
     }
 
     /**
@@ -26,10 +29,9 @@ class PostsController extends Controller
      */
     public function index()
     {
-        //
         //$posts=Post::orderBy('title', 'desc')->get();
         //$posts=DB::select('SELECT * FROM posts');
-        $posts=Post::orderBy('created_at', 'desc')->paginate(11);
+        $posts = Post::orderBy('created_at', 'desc')->paginate(11);
         return view('posts.index')->with('posts', $posts);
     }
 
@@ -40,8 +42,13 @@ class PostsController extends Controller
      */
     public function create()
     {
-        //
-        return view('posts.create');
+        $user = User::find(auth()->id());
+        if ($user->hasRole('poster')) {
+            return view('posts.create');
+        }
+        else {
+            abort(404);
+        }
     }
 
     /**
@@ -58,27 +65,25 @@ class PostsController extends Controller
             'cover_image' => 'image|nullable|max:1999'
         ]);
         //Handle image
-        if($request->hasFile('cover_image')){
-            $complete_filename =$request->file('cover_image')->getClientOriginalName();
-            $filename=pathinfo($complete_filename, PATHINFO_FILENAME);
-            $extension=$request->file('cover_image')->getClientOriginalExtension();
-            $filenNameToStore=$filename.'_'.time().'.'.$extension;
+        if($request->hasFile('cover_image')) {
+            $complete_filename = $request->file('cover_image')->getClientOriginalName();
+            $filename = pathinfo($complete_filename, PATHINFO_FILENAME);
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $filenNameToStore = $filename.'_'.time().'.'.$extension;
             //Image upload
-            $path=$request->file('cover_image')->storeAs('public/cover_images', $filenNameToStore);
-        }
-        else {
-            $filenNameToStore='noimage.jpg';
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $filenNameToStore);
+        } else {
+            $filenNameToStore = 'noimage.jpg';
         }
 
         //Create Post
         $post=new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
-        $post->user_id=auth()->user()->id;
-        $post->cover_image=$filenNameToStore;
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $filenNameToStore;
         $post->save();
         return redirect('/home')->with('success', 'Post Created');
-
     }
 
 
@@ -104,7 +109,7 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
-        if(auth()->user()->id != $post->user_id) {
+        if (auth()->user()->id != $post->user_id) {
             return redirect('/posts')->with('error','Permission Denied');
         }
         return view('posts.edit')->with('post',$post);
@@ -126,7 +131,7 @@ class PostsController extends Controller
         $post=Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
-        if($request->hasFile('cover_image')){
+        if ($request->hasFile('cover_image')) {
             $complete_filename = $request->file('cover_image')->getClientOriginalName();
             $filename = pathinfo($complete_filename, PATHINFO_FILENAME);
             $extension = $request->file('cover_image')->getClientOriginalExtension();
@@ -138,9 +143,8 @@ class PostsController extends Controller
         $post->save();
         return redirect('/posts')->with('success', 'Post Updated');
     }
-
-    /**
-     * Remove the specified resource from storage.
+    /*
+    * Remove the specified resource from storage.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
@@ -148,13 +152,15 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        if(auth()->user()->id != $post->user_id) {
-            return redirect('/posts')->with('error','Permission Denied');
+        if (!(empty($post))) {
+            if (auth()->user()->id != $post->user_id) {
+                return redirect('/posts')->with('error', 'Permission Denied');
+            }
+            if ($post->cover_image != 'noimage.jpg') {
+                Storage::delete('public/cover_images/' . $post->cover_image);
+            }
+            $post->delete();
         }
-        if($post->cover_image!='noimage.jpg') {
-            Storage::delete('public/cover_images/'.$post->cover_image);
-        }
-        $post->delete();
         return redirect('/home')->with('error', 'Post Removed');
     }
 }
